requirements.txt
asgiref==3.9.1
Django==5.2.4
django-extensions==4.1
django-htmx==1.23.2
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
django-ckeditor==6.7.3
